+++ src/store/config.ts
@@ -1,0 +1,12 @@
+import { configureStore } from '@reduxjs/toolkit';
+
+import counterReducer from '~/store/reducers/counterSlice';
+
+export const store = configureStore({
+  reducer: {
+    counter: counterReducer,
+  },
+});
+
+export type RootState = ReturnType<typeof store.getState>;
+export type AppDispatch = typeof store.dispatch;
