+++ src/store/index.ts
@@ -1,12 +1,2 @@
-import { configureStore } from '@reduxjs/toolkit';
-
-import counterReducer from '~/store/reducers/counterSlice';
-
-export const store = configureStore({
-  reducer: {
-    counter: counterReducer,
-  },
-});
-
-export type RootState = ReturnType<typeof store.getState>;
-export type AppDispatch = typeof store.dispatch;
+export * from './config';
+export * from './utils';
