+++ src/store/utils.ts
@@ -1,5 +1,5 @@
 import type { TypedUseSelectorHook } from 'react-redux';
-import type { RootState, AppDispatch } from '~/store';
+import type { RootState, AppDispatch } from '~/store/config';
 
 import { useDispatch, useSelector } from 'react-redux';
